Refactor optimal control demos
