+++ demos/documented/optimal_control/box_constraints/demo_box_constraints.py
@@ -1,25 +1,57 @@
-# Copyright (C) 2020-2022 Sebastian Blauth
-#
-# This file is part of cashocs.
-#
-# cashocs is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# cashocs is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with cashocs.  If not, see <https://www.gnu.org/licenses/>.
+# ---
+# jupyter:
+#   jupytext:
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.14.4
+# ---
 
-"""For the documentation of this demo see
-https://cashocs.readthedocs.io/en/latest/demos/optimal_control/doc_box_constraints.html.
+# (demo_box_constraints)=
+# # Control Constraints
+#
+# ## Problem Formulation
+#
+# In this demo, we take a deeper look at how control constraints can be treated in
+# cashocs. To do so, we investigate the same problem as in {ref}`demo_poisson`, but
+# now with the addition of box constraints for the control variable. This problem
+# reads
+#
+# $$
+# \begin{align}
+#     &\min\; J(y,u) = \frac{1}{2} \int_{\Omega} \left( y - y_d \right)^2 \text{ d}x
+#     + \frac{\alpha}{2} \int_{\Omega} u^2 \text{ d}x \\
+#     &\text{ subject to } \qquad
+#     \begin{alignedat}[t]{2}
+#         -\Delta y &= u \quad &&\text{ in } \Omega,\\
+#         y &= 0 \quad &&\text{ on } \Gamma, \\
+#         u_a \leq u &\leq u_b \quad &&\text{ in } \Omega
+#     \end{alignedat}
+# \end{align}
+# $$
+#
+# (see, e.g., [Tröltzsch - Optimal Control of Partial Differential Equations](
+# https://doi.org/10.1090/gsm/112) or [Hinze, Pinnau, Ulbrich, and Ulbrich -
+# Optimization with PDE constraints](https://doi.org/10.1007/978-1-4020-8839-1).
+#
+# Here, the functions $u_a$ and $u_b$ are $L^\infty(\Omega)$ functions. As before, we
+# consider as domain the unit square, i.e., $\Omega = (0, 1)^2$.
+#
+# ## Implementation
+#
+# The complete python code can be found in the file {download}`demo_box_constraints.py
+# </../../demos/documented/optimal_control/box_constraints/demo_box_constraints.py>`
+# and the corresponding config can be found in {download}`config.ini
+# </../../demos/documented/optimal_control/box_constraints/config.ini>`.
+#
+# ### Initialization
+#
+# The beginning of the script is completely identical to the
+# one of {ref}`previous example <demo_poisson>`, so we only restate the corresponding
+# code in the following
 
-"""
-
+# +
 from fenics import *
 
 import cashocs
@@ -42,26 +74,70 @@ alpha = 1e-6
 J = cashocs.IntegralFunctional(
     Constant(0.5) * (y - y_d) * (y - y_d) * dx + Constant(0.5 * alpha) * u * u * dx
 )
+# -
+
+# ### Definition of the Control Constraints
+#
+# Here, we have nearly everything at hand to define the optimal control problem, the
+# only missing ingredient are the box constraints, which we define now. For the purposes
+# of this example, we consider a linear (in the x-direction) corridor for these
+# constraints, as it highlights the capabilities of cashocs. Hence, we define the lower
+# and upper bounds via
 
 u_a = interpolate(Expression("50*(x[0]-1)", degree=1), V)
 u_b = interpolate(Expression("50*x[0]", degree=1), V)
 
+# which just corresponds to two functions, generated from {py:class}`fenics.Expression`
+# objects via {py:func}`fenics.interpolate`. These are then put into the list `cc`,
+# which models the control constraints, i.e.,
+
 cc = [u_a, u_b]
+
+# ::::{note}
+# As an alternative way of specifying the box constraints, one can also use regular
+# float or int objects, in case that they are constant. For example, the constraint that
+# we only want to consider positive value for u, i.e., $0 \leq u \leq +\infty$ can be
+# realized via
+#
+# :::python
+# u_a = 0
+# u_b = float('inf')
+# cc = [u_a, u_b]
+# :::
+#
+# and completely analogous with `float('-inf')` for no constraint on the lower bound.
+# Moreover, note that the specification of using either constant `float` values and
+# {py:class}`fenics.Function` objects can be mixed arbitrarily, so that one can, e.g.,
+# specify a constant value for the upper boundary and use a {py:class}`fenics.Function`
+# on the lower one.
+# ::::
+#
+# ### Setup of the optimization problem and its solution
+#
+# Now, we can set up the optimal control problem as we did before, using the additional
+# keyword argument `control_constraints` into which we put the list `cc`, and then solve
+# it via the {py:meth}`ocp.solve() <cashocs.OptimalControlProblem.solve>` method
 
 ocp = cashocs.OptimalControlProblem(
     e, bcs, J, y, u, p, config=config, control_constraints=cc
 )
 ocp.solve()
 
+# To check that the box constraints are actually satisfied by our solution, we perform
+# an assertion
+
+# +
 import numpy as np
 
 assert np.alltrue(u_a.vector()[:] <= u.vector()[:]) and np.alltrue(
     u.vector()[:] <= u_b.vector()[:]
 )
+# -
 
+# which shows that they are indeed satisfied. The visualization is carried out
+# analogously to before, via
 
-### Post Processing
-
+# +
 import matplotlib.pyplot as plt
 
 plt.figure(figsize=(15, 5))
@@ -83,3 +159,7 @@ plt.title("Desired state y_d")
 
 plt.tight_layout()
 # plt.savefig('./img_box_constraints.png', dpi=150, bbox_inches='tight')
+# -
+
+# and should yield the following output
+# ![](/../../demos/documented/optimal_control/box_constraints/img_box_constraints.png)
