+++ demos/documented/optimal_control/pre_post_callbacks/img_pre_post_callbacks.png
