+++ docs/source/user/demos/optimal_control/index.rst
@@ -10,19 +10,19 @@ be treated with cashocs.
 
    demo_poisson.md
    doc_config
-   doc_box_constraints
-   doc_neumann_control
-   doc_multiple_variables
-   doc_monolithic_problems
-   doc_picard_iteration
-   doc_stokes
-   doc_heat_equation
-   doc_nonlinear_pdes
-   doc_dirichlet_control
-   doc_iterative_solvers
-   doc_state_constraints
-   doc_sparse_control
-   doc_scalar_control_tracking
-   doc_constraints
-   doc_control_boundary_conditions
-   doc_pre_post_hooks
+   demo_box_constraints.md
+   demo_neumann_control.md
+   demo_multiple_variables.md
+   demo_monolithic_problems.md
+   demo_picard_iteration.md
+   demo_stokes.md
+   demo_heat_equation.md
+   demo_nonlinear_pdes.md
+   demo_dirichlet_control.md
+   demo_iterative_solvers.md
+   demo_state_constraints.md
+   demo_sparse_control.md
+   demo_scalar_control_tracking.md
+   demo_constraints.md
+   demo_control_boundary_conditions.md
+   demo_pre_post_callbacks.md
